\name{npsmReg2-package}
\alias{npsmReg2-package}
\alias{npsmReg2}
\docType{package}
\title{
Chapter 7 Regression II, Kloke and McKean (2014)
}
\description{
Package provides functions discussed in Chapter 7 Regression II, Kloke and McKean (2014).
This includes HBR estimates, robust diagnostics, weighted regression, a Hogg-type adaptive procedure,
nonlinear rank-based fits, and rank-based time series (autoregressive).
}
\details{
\tabular{ll}{
Package: \tab npsmReg2\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-01\cr
License: \tab GPL\cr
}
}
\author{
John Kloke, Joesph McKean

Maintainer: John Kloke <kloke@biostat.wisc.edu>
}
\references{
Kloke, J. and McKean, J.W. (2014), \emph{Nonparametric statistical methods using R}, Boca Raton, FL: Chapman-Hall.

Hettmansperger, T.P. and McKean J.W. (2011), \emph{Robust Nonparametric Statistical Methods, 2nd ed.}, New York: Chapman-Hall.
}
