\name{summary.Rnl}
\alias{summary.Rnl}
\title{Summarize Rank-Based Nonlinear Model Fits }
\description{ Provides a summary similar to the rank-based lonear model fit.  }
\usage{
\method{summary}{wilnl}(object,...)
}
\arguments{
  \item{object}{an object of class 'wilnl', usually, a result of a call to 'wilnl'}
\item{\dots}{ additional arguments }
}
\author{John Kloke \email{kloke@biostat.wisc.edu}}
